report_service/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Marshmallow==0.15.0
marshmallow-sqlalchemy==0.29.0
Flask-CORS==4.0.0
Flask-JWT-Extended==4.5.2
requests==2.31.0
python-dotenv==1.0.0
